requirements.txt
torch~=2.5.1
fastapi~=0.111.0
Pillow~=9.5.0
torchvision~=0.20.1
opencv-python~=4.10.0.84
numpy~=1.26.4
tqdm~=4.66.5
scikit-image~=0.24.0